requirements.txt
langchain-core
langchain-google-genai
google-generativeai
neo4j
python-dotenv